requirements.txt
click
pymongo
python-dotenv